requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
betterlogging==0.2.1
bs4==0.0.1
certifi==2023.7.22
cfscrape==2.1.1
charset-normalizer==3.3.2
cloudscraper==1.2.71
cssselect==1.2.0
dexscreener==0.3
environs==9.5.0
etherscan-python==2.1.0
exceptiongroup==1.1.3
fake-useragent==1.3.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
numpy==1.26.1
outcome==1.3.0.post0
packaging==23.2
parse==1.19.1
py-etherscan-api==0.8.0
pydantic==2.3.0
pydantic_core==2.6.3
pyee==8.2.2
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.0
redis==5.0.1
render-html==1.0.1
requests==2.31.0
requests-html==0.10.0
requests-toolbelt==1.0.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
undetected-chromedriver==3.5.3
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
